rename: 피쳐 구분을 위해 wearable 패키지로 리팩토링
