+++ WearOS/app/src/main/java/com/sandy/logisync/wearable/MessagePath.kt
@@ -1,4 +1,4 @@
-package com.sandy.logisync.service
+package com.sandy.logisync.wearable
 
 enum class MessagePath(val path: String) {
     TEST("/test"),
