+++ WearOS/app/src/main/java/com/sandy/logisync/wearable/TranscriptionPath.kt
@@ -1,4 +1,4 @@
-package com.sandy.logisync.service
+package com.sandy.logisync.wearable
 
 enum class TranscriptionPath(val path: String) {
     SEND_LOGIN_RESPONSE("/login_transcription"),
